requirements.txt
click==8.0.4
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.1
progressbar33==2.4
rdkit==2009.Q1-1
scipy==1.8.0
six==1.16.0
tabulate==0.8.9
torch==1.11.0+cu113
